fix: centralized workflow queries
